Split AsnReader and AsnWriter out to distinct files due to tooling issues.
