+++ src/Common/src/System/Security/Cryptography/AsnReader.cs
@@ -1,0 +1,2773 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Buffers;
+using System.Diagnostics;
+using System.Numerics;
+using System.Runtime.InteropServices;
+using System.Text;
+
+namespace System.Security.Cryptography.Asn1
+{
+    internal class AsnReader
+    {
+        // T-REC-X.690-201508 sec 9.2
+        internal const int MaxCERSegmentSize = 1000;
+
+        // T-REC-X.690-201508 sec 8.1.5 says only 0000 is legal.
+        private const int EndOfContentsEncodedLength = 2;
+
+        private ReadOnlyMemory<byte> _data;
+        private readonly AsnEncodingRules _ruleSet;
+
+        public bool HasData => !_data.IsEmpty;
+
+        public AsnReader(ReadOnlyMemory<byte> data, AsnEncodingRules ruleSet)
+        {
+            CheckEncodingRules(ruleSet);
+
+            _data = data;
+            _ruleSet = ruleSet;
+        }
+
+        public static bool TryPeekTag(ReadOnlySpan<byte> source, out Asn1Tag tag, out int bytesRead)
+        {
+            return Asn1Tag.TryParse(source, out tag, out bytesRead);
+        }
+
+        public Asn1Tag PeekTag()
+        {
+            if (TryPeekTag(_data.Span, out Asn1Tag tag, out int bytesRead))
+            {
+                return tag;
+            }
+
+            throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+        }
+
+        private static bool TryReadLength(
+            ReadOnlySpan<byte> source,
+            AsnEncodingRules ruleSet,
+            out int? length,
+            out int bytesRead)
+        {
+            length = null;
+            bytesRead = 0;
+
+            CheckEncodingRules(ruleSet);
+
+            if (source.IsEmpty)
+            {
+                return false;
+            }
+
+            // T-REC-X.690-201508 sec 8.1.3
+
+            byte lengthOrLengthLength = source[bytesRead];
+            bytesRead++;
+            const byte MultiByteMarker = 0x80;
+
+            // 0x00-0x7F are direct length values.
+            // 0x80 is BER/CER indefinite length.
+            // 0x81-0xFE says that the length takes the next 1-126 bytes.
+            // 0xFF is forbidden.
+            if (lengthOrLengthLength == MultiByteMarker)
+            {
+                // T-REC-X.690-201508 sec 10.1 (DER: Length forms)
+                if (ruleSet == AsnEncodingRules.DER)
+                {
+                    bytesRead = 0;
+                    return false;
+                }
+
+                // Null length == indefinite.
+                return true;
+            }
+
+            if (lengthOrLengthLength < MultiByteMarker)
+            {
+                length = lengthOrLengthLength;
+                return true;
+            }
+
+            if (lengthOrLengthLength == 0xFF)
+            {
+                bytesRead = 0;
+                return false;
+            }
+
+            byte lengthLength = (byte)(lengthOrLengthLength & ~MultiByteMarker);
+
+            // +1 for lengthOrLengthLength
+            if (lengthLength + 1 > source.Length)
+            {
+                bytesRead = 0;
+                return false;
+            }
+
+            // T-REC-X.690-201508 sec 9.1 (CER: Length forms)
+            // T-REC-X.690-201508 sec 10.1 (DER: Length forms)
+            bool minimalRepresentation =
+                ruleSet == AsnEncodingRules.DER || ruleSet == AsnEncodingRules.CER;
+
+            // The ITU-T specifications tecnically allow lengths up to ((2^128) - 1), but
+            // since Span's length is a signed Int32 we're limited to identifying memory
+            // that is within ((2^31) - 1) bytes of the tag start.
+            if (minimalRepresentation && lengthLength > sizeof(int))
+            {
+                bytesRead = 0;
+                return false;
+            }
+
+            uint parsedLength = 0;
+
+            for (int i = 0; i < lengthLength; i++)
+            {
+                byte current = source[bytesRead];
+                bytesRead++;
+
+                if (parsedLength == 0)
+                {
+                    if (minimalRepresentation && current == 0)
+                    {
+                        bytesRead = 0;
+                        return false;
+                    }
+
+                    if (!minimalRepresentation && current != 0)
+                    {
+                        // Under BER rules we could have had padding zeros, so
+                        // once the first data bits come in check that we fit within
+                        // sizeof(int) due to Span bounds.
+
+                        if (lengthLength - i > sizeof(int))
+                        {
+                            bytesRead = 0;
+                            return false;
+                        }
+                    }
+                }
+
+                parsedLength <<= 8;
+                parsedLength |= current;
+            }
+
+            // This value cannot be represented as a Span length.
+            if (parsedLength > int.MaxValue)
+            {
+                bytesRead = 0;
+                return false;
+            }
+
+            if (minimalRepresentation && parsedLength < MultiByteMarker)
+            {
+                bytesRead = 0;
+                return false;
+            }
+
+            Debug.Assert(bytesRead > 0);
+            length = (int)parsedLength;
+            return true;
+        }
+
+        internal Asn1Tag ReadTagAndLength(out int? contentsLength, out int bytesRead)
+        {
+            if (TryPeekTag(_data.Span, out Asn1Tag tag, out int tagBytesRead) &&
+                TryReadLength(_data.Slice(tagBytesRead).Span, _ruleSet, out int? length, out int lengthBytesRead))
+            {
+                int allBytesRead = tagBytesRead + lengthBytesRead;
+
+                if (tag.IsConstructed)
+                {
+                    // T-REC-X.690-201508 sec 9.1 (CER: Length forms) says constructed is always indefinite.
+                    if (_ruleSet == AsnEncodingRules.CER && length != null)
+                    {
+                        throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                    }
+                }
+                else if (length == null)
+                {
+                    // T-REC-X.690-201508 sec 8.1.3.2 says primitive encodings must use a definite form.
+                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                }
+
+                bytesRead = allBytesRead;
+                contentsLength = length;
+                return tag;
+            }
+
+            throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+        }
+
+        private static void ValidateEndOfContents(Asn1Tag tag, int? length, int headerLength)
+        {
+            // T-REC-X.690-201508 sec 8.1.5 excludes the BER 8100 length form for 0.
+            if (tag.IsConstructed || length != 0 || headerLength != EndOfContentsEncodedLength)
+            {
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+        }
+
+        private static ReadOnlyMemory<byte> SeekEndOfContents(
+            ReadOnlyMemory<byte> source,
+            AsnEncodingRules ruleSet,
+            int initialSliceOffset = 0)
+        {
+            ReadOnlyMemory<byte> cur = source.Slice(initialSliceOffset);
+            int totalLen = 0;
+
+            while (!cur.IsEmpty)
+            {
+                AsnReader reader = new AsnReader(cur, ruleSet);
+                Asn1Tag tag = reader.ReadTagAndLength(out int ? length, out int bytesRead);
+                ReadOnlyMemory<byte> nestedContents = reader.PeekContentBytes();
+
+                int localLen = bytesRead + nestedContents.Length;
+
+                if (tag == Asn1Tag.EndOfContents)
+                {
+                    ValidateEndOfContents(tag, length, bytesRead);
+
+                    return source.Slice(0, totalLen + initialSliceOffset);
+                }
+
+                // If the current value was an indefinite-length-encoded value
+                // then we need to skip over the EOC as well.  But we didn't want to
+                // include it as part of the content span.
+                //
+                // T-REC-X.690-201508 sec 8.1.1.1 / 8.1.1.3 indicate that the
+                // End -of-Contents octets are "after" the contents octets, not
+                // "at the end" of them.
+                if (length == null)
+                {
+                    localLen += EndOfContentsEncodedLength;
+                }
+
+                totalLen += localLen;
+                cur = cur.Slice(localLen);
+            }
+
+            throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+        }
+
+        /// <summary>
+        /// Get a ReadOnlyMemory view of the next encoded value without consuming it.
+        /// For indefinite length encodings this includes the End of Contents marker.
+        /// </summary>
+        /// <returns>A ReadOnlyMemory view of the next encoded value.</returns>
+        /// <exception cref="CryptographicException">
+        /// The reader is positioned at a point where the tag or length is invalid
+        /// under the current encoding rules.
+        /// </exception>
+        /// <seealso cref="PeekContentBytes"/>
+        /// <seealso cref="GetEncodedValue"/>
+        public ReadOnlyMemory<byte> PeekEncodedValue()
+        {
+            Asn1Tag tag = ReadTagAndLength(out int ? length, out int bytesRead);
+
+            if (length == null)
+            {
+                ReadOnlyMemory<byte> tagLengthAndContents = SeekEndOfContents(_data, _ruleSet, bytesRead);
+                return Slice(_data, 0, tagLengthAndContents.Length + EndOfContentsEncodedLength);
+            }
+
+            return Slice(_data, 0, bytesRead + length.Value);
+        }
+
+        /// <summary>
+        /// Get a ReadOnlyMemory view of the content octets (bytes) of the next encoded
+        /// value without consuming it.
+        /// </summary>
+        /// <returns>A ReadOnlyMemory view of the contents octets of the next encoded value.</returns>
+        /// <exception cref="CryptographicException">
+        /// The reader is positioned at a point where the tag or length is invalid
+        /// under the current encoding rules.
+        /// </exception>
+        /// <seealso cref="PeekEncodedValue"/>
+        public ReadOnlyMemory<byte> PeekContentBytes()
+        {
+            Asn1Tag tag = ReadTagAndLength(out int ? length, out int bytesRead);
+
+            if (length == null)
+            {
+                return SeekEndOfContents(_data.Slice(bytesRead), _ruleSet);
+            }
+
+            return Slice(_data, bytesRead, length.Value);
+        }
+
+        public void SkipValue()
+        {
+            GetEncodedValue();
+        }
+
+        /// <summary>
+        /// Get a ReadOnlyMemory view of the next encoded value, and move the reader past it.
+        /// For an indefinite length encoding this includes the End of Contents marker.
+        /// </summary>
+        /// <returns>A ReadOnlyMemory view of the next encoded value.</returns>
+        /// <seealso cref="PeekEncodedValue"/>
+        public ReadOnlyMemory<byte> GetEncodedValue()
+        {
+            ReadOnlyMemory<byte> encodedValue = PeekEncodedValue();
+            _data = _data.Slice(encodedValue.Length);
+            return encodedValue;
+        }
+
+        private static bool ReadBooleanValue(
+            ReadOnlySpan<byte> source,
+            AsnEncodingRules ruleSet)
+        {
+            // T-REC-X.690-201508 sec 8.2.1
+            if (source.Length != 1)
+            {
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+
+            byte val = source[0];
+
+            // T-REC-X.690-201508 sec 8.2.2
+            if (val == 0)
+            {
+                return false;
+            }
+
+            // T-REC-X.690-201508 sec 11.1
+            if (val != 0xFF && (ruleSet == AsnEncodingRules.DER || ruleSet == AsnEncodingRules.CER))
+            {
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+
+            return true;
+        }
+
+        public bool ReadBoolean() => ReadBoolean(Asn1Tag.Boolean);
+
+        public bool ReadBoolean(Asn1Tag expectedTag)
+        {
+            Asn1Tag tag = ReadTagAndLength(out int ? length, out int headerLength);
+            CheckExpectedTag(tag, expectedTag, UniversalTagNumber.Boolean);
+
+            // T-REC-X.690-201508 sec 8.2.1
+            if (tag.IsConstructed)
+            {
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+
+            bool value = ReadBooleanValue(
+                Slice(_data, headerLength, length.Value).Span,
+                _ruleSet);
+
+            _data = _data.Slice(headerLength + length.Value);
+            return value;
+        }
+
+        private ReadOnlyMemory<byte> GetIntegerContents(
+            Asn1Tag expectedTag,
+            UniversalTagNumber tagNumber,
+            out int headerLength)
+        {
+            Asn1Tag tag = ReadTagAndLength(out int ? length, out headerLength);
+            CheckExpectedTag(tag, expectedTag, tagNumber);
+
+            // T-REC-X.690-201508 sec 8.3.1
+            if (tag.IsConstructed || length < 1)
+            {
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+
+            // Slice first so that an out of bounds value triggers a CryptographicException.
+            ReadOnlyMemory<byte> contents = Slice(_data, headerLength, length.Value);
+            ReadOnlySpan<byte> contentSpan = contents.Span;
+
+            // T-REC-X.690-201508 sec 8.3.2
+            if (contents.Length > 1)
+            {
+                ushort bigEndianValue = (ushort)(contentSpan[0] << 8 | contentSpan[1]);
+                const ushort RedundancyMask = 0b1111_1111_1000_0000;
+                ushort masked = (ushort)(bigEndianValue & RedundancyMask);
+
+                // If the first 9 bits are all 0 or are all 1, the value is invalid.
+                if (masked == 0 || masked == RedundancyMask)
+                {
+                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                }
+            }
+
+            return contents;
+        }
+
+        public ReadOnlyMemory<byte> GetIntegerBytes() =>
+            GetIntegerBytes(Asn1Tag.Integer);
+
+        public ReadOnlyMemory<byte> GetIntegerBytes(Asn1Tag expectedTag)
+        {
+            ReadOnlyMemory<byte> contents =
+                GetIntegerContents(expectedTag, UniversalTagNumber.Integer, out int headerLength);
+
+            _data = _data.Slice(headerLength + contents.Length);
+            return contents;
+        }
+
+        private bool TryReadSignedInteger(
+            int sizeLimit,
+            Asn1Tag expectedTag,
+            UniversalTagNumber tagNumber,
+            out long value)
+        {
+            Debug.Assert(sizeLimit <= sizeof(long));
+
+            ReadOnlyMemory<byte> contents = GetIntegerContents(expectedTag, tagNumber, out int headerLength);
+
+            if (contents.Length > sizeLimit)
+            {
+                value = 0;
+                return false;
+            }
+
+            ReadOnlySpan<byte> contentSpan = contents.Span;
+
+            bool isNegative = contentSpan[0] >= 0x80;
+            long accum = isNegative ? -1 : 0;
+
+            for (int i = 0; i < contents.Length; i++)
+            {
+                accum <<= 8;
+                accum |= contentSpan[i];
+            }
+
+            _data = _data.Slice(headerLength + contents.Length);
+            value = accum;
+            return true;
+        }
+
+        private bool TryReadUnsignedInteger(
+            int sizeLimit,
+            Asn1Tag expectedTag,
+            UniversalTagNumber tagNumber,
+            out ulong value)
+        {
+            Debug.Assert(sizeLimit <= sizeof(ulong));
+
+            ReadOnlyMemory<byte> contents = GetIntegerContents(expectedTag, tagNumber, out int headerLength);
+            ReadOnlySpan<byte> contentSpan = contents.Span;
+            int contentLength = contents.Length;
+
+            bool isNegative = contentSpan[0] >= 0x80;
+
+            if (isNegative)
+            {
+                // TODO/Review: Should this be "false", an Exception, or not a scenario?
+                value = 0;
+                return false;
+            }
+
+            // Ignore any padding zeros.
+            if (contentSpan.Length > 1 && contentSpan[0] == 0)
+            {
+                contentSpan = contentSpan.Slice(1);
+            }
+
+            if (contentSpan.Length > sizeLimit)
+            {
+                value = 0;
+                return false;
+            }
+
+            ulong accum = 0;
+
+            for (int i = 0; i < contentSpan.Length; i++)
+            {
+                accum <<= 8;
+                accum |= contentSpan[i];
+            }
+
+            _data = _data.Slice(headerLength + contentLength);
+            value = accum;
+            return true;
+        }
+
+        public bool TryReadInt32(out int value) =>
+            TryReadInt32(Asn1Tag.Integer, out value);
+
+        public bool TryReadInt32(Asn1Tag expectedTag, out int value)
+        {
+            if (TryReadSignedInteger(sizeof(int), expectedTag, UniversalTagNumber.Integer, out long longValue))
+            {
+                value = (int)longValue;
+                return true;
+            }
+
+            value = 0;
+            return false;
+        }
+
+        public bool TryReadUInt32(out uint value) =>
+            TryReadUInt32(Asn1Tag.Integer, out value);
+
+        public bool TryReadUInt32(Asn1Tag expectedTag, out uint value)
+        {
+            if (TryReadUnsignedInteger(sizeof(uint), expectedTag, UniversalTagNumber.Integer, out ulong ulongValue))
+            {
+                value = (uint)ulongValue;
+                return true;
+            }
+
+            value = 0;
+            return false;
+        }
+
+        public bool TryReadInt64(out long value) =>
+            TryReadInt64(Asn1Tag.Integer, out value);
+        
+        public bool TryReadInt64(Asn1Tag expectedTag, out long value)
+        {
+            return TryReadSignedInteger(sizeof(long), expectedTag, UniversalTagNumber.Integer, out value);
+        }
+
+        public bool TryReadUInt64(out ulong value) =>
+            TryReadUInt64(Asn1Tag.Integer, out value);
+
+        public bool TryReadUInt64(Asn1Tag expectedTag, out ulong value)
+        {
+            return TryReadUnsignedInteger(sizeof(ulong), expectedTag, UniversalTagNumber.Integer, out value);
+        }
+
+        public bool TryReadInt16(out short value) =>
+            TryReadInt16(Asn1Tag.Integer, out value);
+
+        public bool TryReadInt16(Asn1Tag expectedTag, out short value)
+        {
+            if (TryReadSignedInteger(sizeof(short), expectedTag, UniversalTagNumber.Integer, out long longValue))
+            {
+                value = (short)longValue;
+                return true;
+            }
+
+            value = 0;
+            return false;
+        }
+
+        public bool TryReadUInt16(out ushort value) =>
+            TryReadUInt16(Asn1Tag.Integer, out value);
+
+        public bool TryReadUInt16(Asn1Tag expectedTag, out ushort value)
+        {
+            if (TryReadUnsignedInteger(sizeof(ushort), expectedTag, UniversalTagNumber.Integer, out ulong ulongValue))
+            {
+                value = (ushort)ulongValue;
+                return true;
+            }
+
+            value = 0;
+            return false;
+        }
+
+        public bool TryReadInt8(out sbyte value) =>
+            TryReadInt8(Asn1Tag.Integer, out value);
+
+        public bool TryReadInt8(Asn1Tag expectedTag, out sbyte value)
+        {
+            if (TryReadSignedInteger(sizeof(sbyte), expectedTag, UniversalTagNumber.Integer, out long longValue))
+            {
+                value = (sbyte)longValue;
+                return true;
+            }
+
+            value = 0;
+            return false;
+        }
+
+        public bool TryReadUInt8(out byte value) =>
+            TryReadUInt8(Asn1Tag.Integer, out value);
+
+        public bool TryReadUInt8(Asn1Tag expectedTag, out byte value)
+        {
+            if (TryReadUnsignedInteger(sizeof(byte), expectedTag, UniversalTagNumber.Integer, out ulong ulongValue))
+            {
+                value = (byte)ulongValue;
+                return true;
+            }
+
+            value = 0;
+            return false;
+        }
+
+        private void ParsePrimitiveBitStringContents(
+            ReadOnlyMemory<byte> source,
+            out int unusedBitCount,
+            out ReadOnlyMemory<byte> value,
+            out byte? normalizedLastByte)
+        {
+            // T-REC-X.690-201508 sec 9.2
+            if (_ruleSet == AsnEncodingRules.CER && source.Length > MaxCERSegmentSize)
+            {
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+
+            // T-REC-X.690-201508 sec 8.6.2.3
+            if (source.Length == 0)
+            {
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+
+            ReadOnlySpan<byte> sourceSpan = source.Span;
+            unusedBitCount = sourceSpan[0];
+            normalizedLastByte = null;
+
+            // T-REC-X.690-201508 sec 8.6.2.2
+            if (unusedBitCount > 7)
+            {
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+
+            if (source.Length == 1)
+            {
+                // T-REC-X.690-201508 sec 8.6.2.4
+                if (unusedBitCount > 0)
+                {
+                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                }
+
+                Debug.Assert(unusedBitCount == 0);
+                value = ReadOnlyMemory<byte>.Empty;
+                return;
+            }
+
+            // Build a mask for the bits that are used so the normalized value can be computed
+            //
+            // If 3 bits are "unused" then build a mask for them to check for 0.
+            // 1 << 3 => 0b0000_1000
+            // subtract 1 => 0b0000_0111
+            // Invert that to be the positive mask: 0b111_1000
+            int mask = ~((1 << unusedBitCount) - 1);
+            byte lastByte = sourceSpan[sourceSpan.Length - 1];
+            byte maskedByte = (byte)(lastByte & mask);
+
+            if (maskedByte != lastByte)
+            {
+                // T-REC-X.690-201508 sec 11.2.1
+                if (_ruleSet == AsnEncodingRules.DER || _ruleSet == AsnEncodingRules.CER)
+                {
+                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                }
+
+                normalizedLastByte = maskedByte;
+            }
+
+            value = source.Slice(1);
+        }
+
+        private delegate void BitStringCopyAction(
+            ReadOnlyMemory<byte> value,
+            byte? normalizedLastByte,
+            Span<byte> destination);
+
+        private static void CopyBitStringValue(
+            ReadOnlyMemory<byte> value,
+            byte? normalizedLastByte,
+            Span<byte> destination)
+        {
+            if (value.Length == 0)
+            {
+                return;
+            }
+
+            if (normalizedLastByte == null)
+            {
+                value.Span.CopyTo(destination);
+            }
+            else
+            {
+                int lastIdx = value.Length - 1;
+
+                value.Span.Slice(0, lastIdx).CopyTo(destination);
+                destination[lastIdx] = normalizedLastByte.Value;
+            }
+        }
+
+        private int ProcessConstructedBitString(
+            ReadOnlyMemory<byte> source,
+            ref Span<byte> destination,
+            BitStringCopyAction copyAction,
+            bool isIndefinite,
+            ref int lastUnusedBitCount,
+            ref int lastSegmentLength,
+            out int bytesRead)
+        {
+            if (source.IsEmpty)
+            {
+                if (isIndefinite)
+                {
+                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                }
+
+                bytesRead = 0;
+                return 0;
+            }
+
+            int totalRead = 0;
+            int totalContent = 0;
+            ReadOnlyMemory<byte> cur = source;
+
+            while (!cur.IsEmpty)
+            {
+                AsnReader reader = new AsnReader(cur, _ruleSet);
+                Asn1Tag tag = reader.ReadTagAndLength(out int? length, out int headerLength);
+
+                if (tag.TagClass != TagClass.Universal)
+                {
+                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                }
+
+                totalRead += headerLength;
+
+                if (isIndefinite && tag.TagValue == (int)UniversalTagNumber.EndOfContents)
+                {
+                    ValidateEndOfContents(tag, length, headerLength);
+
+                    bytesRead = totalRead;
+                    return totalContent;
+                }
+
+                if (tag.TagValue != (int)UniversalTagNumber.BitString)
+                {
+                    // T-REC-X.690-201508 sec 8.6.4.1 (in particular, Note 2)
+                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                }
+
+                if (_ruleSet == AsnEncodingRules.CER)
+                {
+                    if (tag.IsConstructed || lastSegmentLength != MaxCERSegmentSize)
+                    {
+                        // T-REC-X.690-201508 sec 9.2
+                        throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                    }
+                }
+
+                cur = cur.Slice(headerLength);
+
+                if (tag.IsConstructed)
+                {
+                    totalContent += ProcessConstructedBitString(
+                        Slice(cur, 0, length),
+                        ref destination,
+                        copyAction,
+                        length == null,
+                        ref lastUnusedBitCount,
+                        ref lastSegmentLength,
+                        out int nestedContentRead);
+
+                    totalRead += nestedContentRead;
+                    cur = cur.Slice(nestedContentRead);
+                }
+                else
+                {
+                    if (lastUnusedBitCount != 0)
+                    {
+                        // T-REC-X.690-201508 sec 8.6.4 (only the last segment can have unused bits)
+                        throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                    }
+
+                    int lengthValue = length.Value;
+
+                    ParsePrimitiveBitStringContents(
+                        Slice(cur, 0, lengthValue),
+                        out lastUnusedBitCount,
+                        out ReadOnlyMemory<byte> value,
+                        out byte? normalizedLastByte);
+
+                    totalRead += lengthValue;
+                    totalContent += value.Length;
+                    lastSegmentLength = lengthValue;
+                    cur = cur.Slice(lengthValue);
+
+                    if (copyAction != null)
+                    {
+                        copyAction(value, normalizedLastByte, destination);
+
+                        destination = destination.Slice(value.Length);
+                    }
+                }
+            }
+
+            if (isIndefinite)
+            {
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+
+            bytesRead = totalRead;
+            return totalContent;
+        }
+
+        private bool TryCopyConstructedBitStringValue(
+            ReadOnlyMemory<byte> source,
+            Span<byte> dest,
+            bool isIndefinite,
+            out int unusedBitCount,
+            out int bytesRead,
+            out int bytesWritten)
+        {
+            int lastUnusedBitCount = 0;
+            int lastSegmentSize = MaxCERSegmentSize;
+
+            Span<byte> tmpDest = dest;
+
+            // Call ProcessConstructedBitString with a null copy-action to get the required byte
+            // count and verify that the data is well-formed before copying into dest.
+            int contentLength = ProcessConstructedBitString(
+                source,
+                ref tmpDest,
+                null,
+                isIndefinite,
+                ref lastUnusedBitCount,
+                ref lastSegmentSize,
+                out int encodedLength);
+
+            // Since the unused bits byte from the segments don't count, only one segment
+            // returns 999 (or less), the second segment bumps the count to 1000, and is legal.
+            //
+            // T-REC-X.690-201508 sec 9.2
+            if (_ruleSet == AsnEncodingRules.CER && contentLength < MaxCERSegmentSize)
+            {
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+
+            if (dest.Length < contentLength)
+            {
+                unusedBitCount = 0;
+                bytesRead = 0;
+                bytesWritten = 0;
+                return false;
+            }
+
+            unusedBitCount = lastUnusedBitCount;
+            tmpDest = dest;
+            lastSegmentSize = MaxCERSegmentSize;
+            lastUnusedBitCount = 0;
+
+            // Now call it again with the method which fixes the normalized byte in the last segment.
+            bytesWritten = ProcessConstructedBitString(
+                source,
+                ref tmpDest,
+                (value, lastByte, destination) => CopyBitStringValue(value, lastByte, destination),
+                isIndefinite,
+                ref lastUnusedBitCount,
+                ref lastSegmentSize,
+                out bytesRead);
+
+            Debug.Assert(unusedBitCount == lastUnusedBitCount);
+            Debug.Assert(bytesWritten == contentLength);
+            Debug.Assert(bytesRead == encodedLength);
+            return true;
+        }
+
+        private bool TryGetPrimitiveBitStringValue(
+            Asn1Tag expectedTag,
+            out Asn1Tag actualTag,
+            out int? contentsLength,
+            out int headerLength,
+            out int unusedBitCount,
+            out ReadOnlyMemory<byte> value,
+            out byte? normalizedLastByte)
+        {
+            actualTag = ReadTagAndLength(out contentsLength, out headerLength);
+            CheckExpectedTag(actualTag, expectedTag, UniversalTagNumber.BitString);
+
+            if (actualTag.IsConstructed)
+            {
+                if (_ruleSet == AsnEncodingRules.DER)
+                {
+                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                }
+
+                unusedBitCount = 0;
+                value = default(ReadOnlyMemory<byte>);
+                normalizedLastByte = null;
+                return false;
+            }
+
+            Debug.Assert(contentsLength.HasValue);
+            ReadOnlyMemory<byte> encodedValue = Slice(_data, headerLength, contentsLength.Value);
+
+            ParsePrimitiveBitStringContents(
+                encodedValue,
+                out unusedBitCount,
+                out value,
+                out normalizedLastByte);
+
+            return true;
+        }
+
+        public bool TryGetPrimitiveBitStringValue(out int unusedBitCount, out ReadOnlyMemory<byte> contents)
+            => TryGetPrimitiveBitStringValue(Asn1Tag.PrimitiveBitString, out unusedBitCount, out contents);
+
+        /// <summary>
+        /// Gets a ReadOnlyMemory view over the data value portion of the contents of a bit string.
+        /// </summary>
+        /// <param name="expectedTag">The expected tag to read</param>
+        /// <param name="unusedBitCount">The encoded value for the number of unused bits.</param>
+        /// <param name="value">The data value portion of the bit string contents.</param>
+        /// <returns>
+        ///   <c>true</c> if the bit string uses a primitive encoding and the "unused" bits have value 0,
+        ///   <c>false</c> otherwise.
+        /// </returns>
+        /// <exception cref="CryptographicException">
+        ///  <ul>
+        ///   <li>No data remains</li>
+        ///   <li>The tag read does not match the expected tag</li>
+        ///   <li>The length is invalid under the chosen encoding rules</li>
+        ///   <li>The unusedBitCount value is out of bounds</li>
+        ///   <li>A CER or DER encoding was chosen and an "unused" bit was set to 1</li>
+        ///   <li>A CER encoding was chosen and the primitive content length exceeds the maximum allowed</li>
+        /// </ul>
+        /// </exception>
+        public bool TryGetPrimitiveBitStringValue(
+            Asn1Tag expectedTag,
+            out int unusedBitCount,
+            out ReadOnlyMemory<byte> value)
+        {
+            bool isPrimitive = TryGetPrimitiveBitStringValue(
+                expectedTag,
+                out Asn1Tag actualTag,
+                out int? contentsLength,
+                out int headerLength,
+                out unusedBitCount,
+                out value,
+                out byte? normalizedLastByte);
+
+            if (isPrimitive)
+            {
+                // A BER reader which encountered a situation where an "unused" bit was not
+                // set to 0.
+                if (normalizedLastByte != null)
+                {
+                    unusedBitCount = 0;
+                    value = default(ReadOnlyMemory<byte>);
+                    return false;
+                }
+
+                // Skip the tag+length (header) and the unused bit count byte (1) and the contents.
+                _data = _data.Slice(headerLength + value.Length + 1);
+            }
+
+            return isPrimitive;
+        }
+
+        public bool TryCopyBitStringBytes(
+            Span<byte> destination,
+            out int unusedBitCount,
+            out int bytesWritten)
+        {
+            return TryCopyBitStringBytes(
+                Asn1Tag.PrimitiveBitString,
+                destination,
+                out unusedBitCount,
+                out bytesWritten);
+        }
+
+        public bool TryCopyBitStringBytes(
+            Asn1Tag expectedTag,
+            Span<byte> destination,
+            out int unusedBitCount,
+            out int bytesWritten)
+        {
+            if (TryGetPrimitiveBitStringValue(
+                expectedTag,
+                out Asn1Tag actualTag,
+                out int? contentsLength,
+                out int headerLength,
+                out unusedBitCount,
+                out ReadOnlyMemory<byte> value,
+                out byte? normalizedLastByte))
+            {
+                if (value.Length > destination.Length)
+                {
+                    bytesWritten = 0;
+                    unusedBitCount = 0;
+                    return false;
+                }
+
+                CopyBitStringValue(value, normalizedLastByte, destination);
+
+                bytesWritten = value.Length;
+                // contents doesn't include the unusedBitCount value, so add one byte for that.
+                _data = _data.Slice(headerLength + value.Length + 1);
+                return true;
+            }
+
+            Debug.Assert(actualTag.IsConstructed);
+
+            bool read = TryCopyConstructedBitStringValue(
+                Slice(_data, headerLength, contentsLength),
+                destination,
+                contentsLength == null,
+                out unusedBitCount,
+                out int bytesRead,
+                out bytesWritten);
+
+            if (read)
+            {
+                _data = _data.Slice(headerLength + bytesRead);
+            }
+
+            return read;
+        }
+
+        public TFlagsEnum GetNamedBitListValue<TFlagsEnum>() where TFlagsEnum : struct =>
+            GetNamedBitListValue<TFlagsEnum>(Asn1Tag.PrimitiveBitString);
+
+        public TFlagsEnum GetNamedBitListValue<TFlagsEnum>(Asn1Tag expectedTag) where TFlagsEnum : struct
+        {
+            Type tFlagsEnum = typeof(TFlagsEnum);
+
+            return (TFlagsEnum)Enum.ToObject(tFlagsEnum, GetNamedBitListValue(expectedTag, tFlagsEnum));
+        }
+
+        public Enum GetNamedBitListValue(Type tFlagsEnum) =>
+            GetNamedBitListValue(Asn1Tag.PrimitiveBitString, tFlagsEnum);
+
+        public Enum GetNamedBitListValue(Asn1Tag expectedTag, Type tFlagsEnum)
+        {
+            // This will throw an ArgumentException if TEnum isn't an enum type,
+            // so we don't need to validate it.
+            Type backingType = tFlagsEnum.GetEnumUnderlyingType();
+
+            if (!tFlagsEnum.IsDefined(typeof(FlagsAttribute), false))
+            {
+                throw new ArgumentException(
+                    SR.Cryptography_Asn_NamedBitListRequiresFlagsEnum,
+                    nameof(tFlagsEnum));
+            }
+
+            int sizeLimit = Marshal.SizeOf(backingType);
+            Span<byte> stackSpan = stackalloc byte[sizeLimit];
+            ReadOnlyMemory<byte> saveData = _data;
+
+            if (!TryCopyBitStringBytes(expectedTag, stackSpan, out int unusedBitCount, out int bytesWritten))
+            {
+                throw new CryptographicException(
+                    SR.Format(SR.Cryptography_Asn_NamedBitListValueTooBig, tFlagsEnum.Name));
+            }
+
+            if (bytesWritten == 0)
+            {
+                // The mode isn't relevant, zero is always zero.
+                return (Enum)Enum.ToObject(tFlagsEnum, 0);
+            }
+
+            ReadOnlySpan<byte> valueSpan = stackSpan.Slice(0, bytesWritten);
+
+            // Now that the 0-bounds check is out of the way:
+            // 
+            // T-REC-X.690-201508 sec 11.2.2
+            if (_ruleSet == AsnEncodingRules.DER ||
+                _ruleSet == AsnEncodingRules.CER)
+            {
+                byte lastByte = valueSpan[bytesWritten - 1];
+
+                // No unused bits tests 0x01, 1 is 0x02, 2 is 0x04, etc.
+                // We already know that TryCopyBitStringBytes checked that the
+                // declared unused bits were 0, this checks that the last "used" bit
+                // isn't also zero.
+                byte testBit = (byte)(1 << unusedBitCount);
+
+                if ((lastByte & testBit) == 0)
+                {
+                    _data = saveData;
+                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                }
+            }
+
+            // Consider a NamedBitList defined as
+            //
+            //   SomeList ::= BIT STRING {
+            //     a(0), b(1), c(2), d(3), e(4), f(5), g(6), h(7), i(8), j(9), k(10)
+            //   }
+            //
+            // The BIT STRING encoding of (a | j) is
+            //   unusedBitCount = 6,
+            //   contents: 0x80 0x40  (0b10000000_01000000)
+            //
+            // A the C# exposure of this structure we adhere to is
+            //
+            // [Flags]
+            // enum SomeList
+            // {
+            //     A = 1,
+            //     B = 1 << 1,
+            //     C = 1 << 2,
+            //     ...
+            // }
+            //
+            // Which happens to be exactly backwards from how the bits are encoded, but the complexity
+            // only needs to live here.
+            return (Enum)Enum.ToObject(tFlagsEnum, InterpretNamedBitListReversed(valueSpan));
+        }
+
+        private static long InterpretNamedBitListReversed(ReadOnlySpan<byte> valueSpan)
+        {
+            Debug.Assert(valueSpan.Length <= sizeof(long));
+
+            long accum = 0;
+            long currentBitValue = 1;
+
+            for (int byteIdx = 0; byteIdx < valueSpan.Length; byteIdx++)
+            {
+                byte byteVal = valueSpan[byteIdx];
+
+                for (int bitIndex = 7; bitIndex >= 0; bitIndex--)
+                {
+                    int test = 1 << bitIndex;
+
+                    if ((byteVal & test) != 0)
+                    {
+                        accum |= currentBitValue;
+                    }
+
+                    currentBitValue <<= 1;
+                }
+            }
+
+            return accum;
+        }
+
+        public ReadOnlyMemory<byte> GetEnumeratedBytes() =>
+            GetEnumeratedBytes(Asn1Tag.Enumerated);
+
+        public ReadOnlyMemory<byte> GetEnumeratedBytes(Asn1Tag expectedTag)
+        {
+            // T-REC-X.690-201508 sec 8.4 says the contents are the same as for integers.
+            ReadOnlyMemory<byte> contents =
+                GetIntegerContents(expectedTag, UniversalTagNumber.Enumerated, out int headerLength);
+
+            _data = _data.Slice(headerLength + contents.Length);
+            return contents;
+        }
+
+        public TEnum GetEnumeratedValue<TEnum>() where TEnum : struct
+        {
+            Type tEnum = typeof(TEnum);
+
+            return (TEnum)Enum.ToObject(tEnum, GetEnumeratedValue(tEnum));
+        }
+
+        public TEnum GetEnumeratedValue<TEnum>(Asn1Tag expectedTag) where TEnum : struct
+        {
+            Type tEnum = typeof(TEnum);
+
+            return (TEnum)Enum.ToObject(tEnum, GetEnumeratedValue(expectedTag, tEnum));
+        }
+
+        public Enum GetEnumeratedValue(Type tEnum) =>
+            GetEnumeratedValue(Asn1Tag.Enumerated, tEnum);
+
+        public Enum GetEnumeratedValue(Asn1Tag expectedTag, Type tEnum)
+        {
+            const UniversalTagNumber tagNumber = UniversalTagNumber.Enumerated;
+            
+            // This will throw an ArgumentException if TEnum isn't an enum type,
+            // so we don't need to validate it.
+            Type backingType = tEnum.GetEnumUnderlyingType();
+
+            if (tEnum.IsDefined(typeof(FlagsAttribute), false))
+            {
+                throw new ArgumentException(
+                    SR.Cryptography_Asn_EnumeratedValueRequiresNonFlagsEnum,
+                    nameof(tEnum));
+            }
+
+            // T-REC-X.690-201508 sec 8.4 says the contents are the same as for integers.
+            int sizeLimit = Marshal.SizeOf(backingType);
+
+            if (backingType == typeof(int) ||
+                backingType == typeof(long) ||
+                backingType == typeof(short) ||
+                backingType == typeof(sbyte))
+            {
+                if (!TryReadSignedInteger(sizeLimit, expectedTag, tagNumber, out long value))
+                {
+                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                }
+
+                return (Enum)Enum.ToObject(tEnum, value);
+            }
+
+            if (backingType == typeof(uint) ||
+                backingType == typeof(ulong) ||
+                backingType == typeof(ushort) ||
+                backingType == typeof(byte))
+            {
+                if (!TryReadUnsignedInteger(sizeLimit, expectedTag, tagNumber, out ulong value))
+                {
+                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                }
+
+                return (Enum)Enum.ToObject(tEnum, value);
+            }
+
+            Debug.Fail($"No handler for type {backingType.Name}");
+            throw new CryptographicException();
+        }
+
+        private bool TryGetOctetStringBytes(
+            Asn1Tag expectedTag,
+            out ReadOnlyMemory<byte> contents,
+            out int headerLength,
+            UniversalTagNumber universalTagNumber = UniversalTagNumber.OctetString)
+        {
+            Asn1Tag tag = ReadTagAndLength(out int ? length, out headerLength);
+            CheckExpectedTag(tag, expectedTag, universalTagNumber);
+
+            if (tag.IsConstructed)
+            {
+                if (_ruleSet == AsnEncodingRules.DER)
+                {
+                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                }
+
+                contents = default(ReadOnlyMemory<byte>);
+                return false;
+            }
+
+            Debug.Assert(length.HasValue);
+            ReadOnlyMemory<byte> encodedValue = Slice(_data, headerLength, length.Value);
+
+            if (_ruleSet == AsnEncodingRules.CER && encodedValue.Length > MaxCERSegmentSize)
+            {
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+
+            contents = encodedValue;
+            return true;
+        }
+
+        private bool TryGetOctetStringBytes(
+            Asn1Tag expectedTag,
+            UniversalTagNumber universalTagNumber,
+            out ReadOnlyMemory<byte> contents)
+        {
+            if (TryGetOctetStringBytes(expectedTag, out contents, out int headerLength, universalTagNumber))
+            {
+                _data = _data.Slice(headerLength + contents.Length);
+                return true;
+            }
+
+            return false;
+        }
+
+        public bool TryGetOctetStringBytes(out ReadOnlyMemory<byte> contents) =>
+            TryGetOctetStringBytes(Asn1Tag.PrimitiveOctetString, out contents);
+
+        /// <summary>
+        /// Gets the source data for an OctetString under a primitive encoding.
+        /// </summary>
+        /// <param name="expectedTag">The expected tag value</param>
+        /// <param name="contents">The content bytes for the OctetString payload.</param>
+        /// <returns>
+        ///   <c>true</c> if the octet string uses a primitive encoding, <c>false</c> otherwise.
+        /// </returns>
+        /// <exception cref="CryptographicException">
+        ///  <ul>
+        ///   <li>No data remains</li>
+        ///   <li>The tag read did not match the expected tag</li>
+        ///   <li>The length is invalid under the chosen encoding rules</li>
+        ///   <li>A CER encoding was chosen and the primitive content length exceeds the maximum allowed</li>
+        /// </ul>
+        /// </exception>
+        public bool TryGetOctetStringBytes(Asn1Tag expectedTag, out ReadOnlyMemory<byte> contents)
+        {
+            return TryGetOctetStringBytes(expectedTag, UniversalTagNumber.OctetString, out contents);
+        }
+
+        private static int CopyConstructedOctetString(
+            ReadOnlyMemory<byte> source,
+            ref Span<byte> destination,
+            bool write,
+            AsnEncodingRules ruleSet,
+            bool isIndefinite,
+            ref int lastSegmentLength,
+            out int bytesRead)
+        {
+            if (source.IsEmpty)
+            {
+                if (isIndefinite)
+                {
+                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                }
+
+                bytesRead = 0;
+                return 0;
+            }
+
+            int totalRead = 0;
+            int totalContent = 0;
+            ReadOnlyMemory<byte> cur = source;
+
+            while (!cur.IsEmpty)
+            {
+                AsnReader reader = new AsnReader(cur, ruleSet);
+                Asn1Tag tag = reader.ReadTagAndLength(out int? length, out int headerLength);
+                
+                if (tag.TagClass != TagClass.Universal)
+                {
+                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                }
+
+                totalRead += headerLength;
+
+                if (isIndefinite && tag.TagValue == (int)UniversalTagNumber.EndOfContents)
+                {
+                    ValidateEndOfContents(tag, length, headerLength);
+
+                    bytesRead = totalRead;
+                    return totalContent;
+                }
+
+                if (tag.TagValue != (int)UniversalTagNumber.OctetString)
+                {
+                    // T-REC-X.690-201508 sec 8.7.3.2 (in particular, Note 2)
+                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                }
+
+                if (ruleSet == AsnEncodingRules.CER)
+                {
+                    if (tag.IsConstructed || lastSegmentLength != MaxCERSegmentSize)
+                    {
+                        // T-REC-X.690-201508 sec 9.2
+                        throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                    }
+                }
+
+                cur = cur.Slice(headerLength);
+
+                if (length == null)
+                {
+                    Debug.Assert(tag.IsConstructed);
+
+                    totalContent += CopyConstructedOctetString(
+                        cur,
+                        ref destination,
+                        write,
+                        ruleSet,
+                        true,
+                        ref lastSegmentLength,
+                        out int nestedBytesRead);
+
+                    totalRead += nestedBytesRead;
+                    cur = cur.Slice(nestedBytesRead);
+                }
+                else if (tag.IsConstructed)
+                {
+                    totalContent += CopyConstructedOctetString(
+                        Slice(cur, 0, length.Value),
+                        ref destination,
+                        write,
+                        ruleSet,
+                        false,
+                        ref lastSegmentLength,
+                        out int nestedContentRead);
+
+                    totalRead += nestedContentRead;
+                    cur = cur.Slice(nestedContentRead);
+                }
+                else
+                {
+                    int lengthValue = length.Value;
+
+                    // T-REC-X.690-201508 sec 9.2
+                    if (ruleSet == AsnEncodingRules.CER && lengthValue > MaxCERSegmentSize)
+                    {
+                        throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                    }
+
+                    totalRead += lengthValue;
+                    totalContent += lengthValue;
+                    lastSegmentLength = lengthValue;
+
+                    ReadOnlyMemory<byte> segment = Slice(cur, 0, lengthValue);
+                    cur = cur.Slice(lengthValue);
+
+                    if (write)
+                    {
+                        segment.Span.CopyTo(destination);
+                        destination = destination.Slice(lengthValue);
+                    }
+                }
+            }
+
+            if (isIndefinite)
+            {
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+
+            bytesRead = totalRead;
+            return totalContent;
+        }
+
+        private static bool TryCopyConstructedOctetStringValue(
+            ReadOnlyMemory<byte> source,
+            Span<byte> dest,
+            bool write,
+            AsnEncodingRules ruleSet,
+            bool isIndefinite,
+            out int bytesRead,
+            out int bytesWritten)
+        {
+            int lastSegmentSize = MaxCERSegmentSize;
+
+            Span<byte> tmpDest = dest;
+
+            int contentLength = CopyConstructedOctetString(
+                source,
+                ref tmpDest,
+                false,
+                ruleSet,
+                isIndefinite,
+                ref lastSegmentSize,
+                out int encodedLength);
+
+            // T-REC-X.690-201508 sec 9.2
+            if (ruleSet == AsnEncodingRules.CER && contentLength <= MaxCERSegmentSize)
+            {
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+
+            if (!write)
+            {
+                bytesRead = encodedLength;
+                bytesWritten = contentLength;
+                return true;
+            }
+
+            if (dest.Length < contentLength)
+            {
+                bytesRead = 0;
+                bytesWritten = 0;
+                return false;
+            }
+
+            tmpDest = dest;
+            lastSegmentSize = MaxCERSegmentSize;
+
+            bytesWritten = CopyConstructedOctetString(
+                source,
+                ref tmpDest,
+                true,
+                ruleSet,
+                isIndefinite,
+                ref lastSegmentSize,
+                out bytesRead);
+
+            Debug.Assert(bytesWritten == contentLength);
+            Debug.Assert(bytesRead == encodedLength);
+            return true;
+        }
+
+        public bool TryCopyOctetStringBytes(
+            Span<byte> destination,
+            out int bytesWritten)
+        {
+            return TryCopyOctetStringBytes(
+                Asn1Tag.PrimitiveOctetString,
+                destination,
+                out bytesWritten);
+        }
+
+        public bool TryCopyOctetStringBytes(
+            Asn1Tag expectedTag,
+            Span<byte> destination,
+            out int bytesWritten)
+        {
+            if (TryGetOctetStringBytes(
+                expectedTag,
+                out ReadOnlyMemory<byte> contents,
+                out int headerLength))
+            {
+                if (contents.Length > destination.Length)
+                {
+                    bytesWritten = 0;
+                    return false;
+                }
+
+                contents.Span.CopyTo(destination);
+                bytesWritten = contents.Length;
+                _data = _data.Slice(headerLength + contents.Length);
+                return true;
+            }
+
+            Asn1Tag tag = ReadTagAndLength(out int ? length, out headerLength);
+
+            bool copied = TryCopyConstructedOctetStringValue(
+                Slice(_data, headerLength, length),
+                destination,
+                true,
+                _ruleSet,
+                length == null,
+                out int bytesRead,
+                out bytesWritten);
+
+            if (copied)
+            {
+                _data = _data.Slice(headerLength + bytesRead);
+            }
+
+            return copied;
+        }
+
+        public void ReadNull() => ReadNull(Asn1Tag.Null);
+
+        public void ReadNull(Asn1Tag expectedTag)
+        {
+            Asn1Tag tag = ReadTagAndLength(out int ? length, out int headerLength);
+            CheckExpectedTag(tag, expectedTag, UniversalTagNumber.Null);
+
+            // T-REC-X.690-201508 sec 8.8.1
+            // T-REC-X.690-201508 sec 8.8.2
+            if (tag.IsConstructed || length != 0)
+            {
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+
+            _data = _data.Slice(headerLength);
+        }
+        
+        private static BigInteger ReadSubIdentifier(
+            ReadOnlySpan<byte> source,
+            out int bytesRead)
+        {
+            Debug.Assert(source.Length > 0);
+
+            // T-REC-X.690-201508 sec 8.19.2 (last sentence)
+            if (source[0] == 0x80)
+            {
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+
+            BigInteger accum = BigInteger.Zero;
+
+            for (int idx = 0; idx < source.Length; idx++)
+            {
+                byte cur = source[idx];
+
+                accum <<= 7;
+                accum |= (cur & 0x7F);
+
+                // If the high bit isn't set this marks the end of the sub-identifier.
+                if (cur < 0x80)
+                {
+                    bytesRead = idx + 1;
+                    return accum;
+                }
+            }
+
+            throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+        }
+
+        private string ReadObjectIdentifierAsString(Asn1Tag expectedTag, out int totalBytesRead)
+        {
+            Asn1Tag tag = ReadTagAndLength(out int ? length, out int headerLength);
+            CheckExpectedTag(tag, expectedTag, UniversalTagNumber.ObjectIdentifier);
+
+            // T-REC-X.690-201508 sec 8.19.1
+            // T-REC-X.690-201508 sec 8.19.2 says the minimum length is 1
+            if (tag.IsConstructed || length < 1)
+            {
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+
+            ReadOnlyMemory<byte> contentsMemory = Slice(_data, headerLength, length.Value);
+            ReadOnlySpan<byte> contents = contentsMemory.Span;
+
+            BigInteger firstIdentifier = ReadSubIdentifier(contents, out int bytesRead);
+            byte firstArc;
+
+            // T-REC-X.690-201508 sec 8.19.4
+            // The first two subidentifiers (X.Y) are encoded as (X * 40) + Y, because Y is
+            // bounded [0, 39] for X in {0, 1}, and only X in {0, 1, 2} are legal.
+            // So:
+            // * identifier < 40 => X = 0, Y = identifier.
+            // * identifier < 80 => X = 1, Y = identifier - 40.
+            // * else: X = 2, Y = identifier - 80.
+
+            if (firstIdentifier < 40)
+            {
+                firstArc = 0;
+            }
+            else if (firstIdentifier < 80)
+            {
+                firstArc = 1;
+                firstIdentifier -= 40;
+            }
+            else
+            {
+                firstArc = 2;
+                firstIdentifier -= 80;
+            }
+
+            StringBuilder builder = new StringBuilder(contents.Length * 4);
+            builder.Append(firstArc);
+            builder.Append('.');
+            builder.Append(firstIdentifier.ToString());
+
+            contents = contents.Slice(bytesRead);
+
+            while (!contents.IsEmpty)
+            {
+                BigInteger subIdentifier = ReadSubIdentifier(contents, out bytesRead);
+                builder.Append('.');
+                builder.Append(subIdentifier.ToString());
+
+                contents = contents.Slice(bytesRead);
+            }
+
+            totalBytesRead = headerLength + length.Value;
+            return builder.ToString();
+        }
+
+        public string ReadObjectIdentifierAsString() =>
+            ReadObjectIdentifierAsString(Asn1Tag.ObjectIdentifier);
+        
+        public string ReadObjectIdentifierAsString(Asn1Tag expectedTag)
+        {
+            string oidValue = ReadObjectIdentifierAsString(expectedTag, out int bytesRead);
+
+            _data = _data.Slice(bytesRead);
+
+            return oidValue;
+        }
+
+        public Oid ReadObjectIdentifier(bool skipFriendlyName = false) =>
+            ReadObjectIdentifier(Asn1Tag.ObjectIdentifier, skipFriendlyName);
+
+        public Oid ReadObjectIdentifier(Asn1Tag expectedTag, bool skipFriendlyName=false)
+        {
+            string oidValue = ReadObjectIdentifierAsString(expectedTag, out int bytesRead);
+            Oid oid = skipFriendlyName ? new Oid(oidValue, oidValue) : new Oid(oidValue);
+
+            // Don't slice until the return object has been created.
+            _data = _data.Slice(bytesRead);
+
+            return oid;
+        }
+
+        private bool TryCopyCharacterStringBytes(
+            Asn1Tag expectedTag,
+            UniversalTagNumber universalTagNumber,
+            Span<byte> destination,
+            bool write,
+            out int bytesRead,
+            out int bytesWritten)
+        {
+            if (TryGetOctetStringBytes(
+                expectedTag,
+                out ReadOnlyMemory<byte> contents,
+                out int headerLength,
+                universalTagNumber))
+            {
+                bytesWritten = contents.Length;
+
+                if (write)
+                {
+                    if (destination.Length < bytesWritten)
+                    {
+                        bytesWritten = 0;
+                        bytesRead = 0;
+                        return false;
+                    }
+
+                    contents.Span.CopyTo(destination);
+                }
+
+                bytesRead = headerLength + bytesWritten;
+                return true;
+            }
+
+            Asn1Tag tag = ReadTagAndLength(out int ? length, out headerLength);
+
+            bool copied = TryCopyConstructedOctetStringValue(
+                Slice(_data, headerLength, length),
+                destination,
+                write,
+                _ruleSet,
+                length == null,
+                out int contentBytesRead,
+                out bytesWritten);
+
+            bytesRead = headerLength + contentBytesRead;
+            return copied;
+        }
+
+        private static unsafe string GetCharacterString(
+            ReadOnlyMemory<byte> source,
+            Text.Encoding encoding)
+        {
+            fixed (byte* bytePtr = &source.Span.DangerousGetPinnableReference())
+            {
+                try
+                {
+                    return encoding.GetString(bytePtr, source.Length);
+                }
+                catch (DecoderFallbackException e)
+                {
+                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding, e);
+                }
+            }
+        }
+
+        private static unsafe bool TryCopyCharacterString(
+            ReadOnlyMemory<byte> source,
+            Text.Encoding encoding,
+            Span<char> destination,
+            out int charsWritten)
+        {
+            fixed (byte* bytePtr = &source.Span.DangerousGetPinnableReference())
+            fixed (char* charPtr = &destination.DangerousGetPinnableReference())
+            {
+                try
+                {
+                    int charCount = encoding.GetCharCount(bytePtr, source.Length);
+
+                    if (charCount > destination.Length)
+                    {
+                        charsWritten = 0;
+                        return false;
+                    }
+
+                    charsWritten = encoding.GetChars(bytePtr, source.Length, charPtr, destination.Length);
+                    Debug.Assert(charCount == charsWritten);
+                }
+                catch (DecoderFallbackException e)
+                {
+                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding, e);
+                }
+
+                return true;
+            }
+        }
+
+        private string GetCharacterString(
+            Asn1Tag expectedTag,
+            UniversalTagNumber universalTagNumber,
+            Text.Encoding encoding)
+        {
+            if (TryGetOctetStringBytes(
+                expectedTag,
+                out ReadOnlyMemory<byte> contents,
+                out int headerLength,
+                universalTagNumber))
+            {
+                string s = GetCharacterString(contents, encoding);
+
+                _data = _data.Slice(headerLength + contents.Length);
+
+                return s;
+            }
+
+            bool parsed = TryCopyCharacterStringBytes(
+                expectedTag,
+                universalTagNumber,
+                Span<byte>.Empty,
+                false,
+                out int bytesRead,
+                out int bytesWritten);
+
+            Debug.Assert(parsed, "TryCopyCharacterStringBytes returned false in counting mode");
+
+            byte[] rented = ArrayPool<byte>.Shared.Rent(bytesWritten);
+
+            try
+            {
+                if (!TryCopyCharacterStringBytes(expectedTag, universalTagNumber, rented, true, out bytesRead, out bytesWritten))
+                {
+                    Debug.Fail("TryCopyCharacterStringBytes failed with a precomputed size");
+                    throw new CryptographicException();
+                }
+
+                string s = GetCharacterString(
+                    new ReadOnlyMemory<byte>(rented, 0, bytesWritten),
+                    encoding);
+
+                _data = _data.Slice(bytesRead);
+
+                return s;
+            }
+            finally
+            {
+                Array.Clear(rented, 0, bytesWritten);
+                ArrayPool<byte>.Shared.Return(rented);
+            }
+        }
+
+        private bool TryCopyCharacterString(
+            Asn1Tag expectedTag,
+            UniversalTagNumber universalTagNumber,
+            Text.Encoding encoding,
+            Span<char> destination,
+            out int charsWritten)
+        {
+            if (TryGetOctetStringBytes(
+                expectedTag,
+                out ReadOnlyMemory<byte> contents,
+                out int headerLength,
+                universalTagNumber))
+            {
+                bool copied = TryCopyCharacterString(contents, encoding, destination, out charsWritten);
+
+                if (copied)
+                {
+                    _data = _data.Slice(headerLength + contents.Length);
+                }
+
+                return copied;
+            }
+
+            bool parsed = TryCopyCharacterStringBytes(
+                expectedTag,
+                universalTagNumber,
+                Span<byte>.Empty,
+                false,
+                out int bytesRead,
+                out int bytesWritten);
+
+            Debug.Assert(parsed, "TryCopyCharacterStringBytes returned false in counting mode");
+
+            byte[] rented = ArrayPool<byte>.Shared.Rent(bytesWritten);
+
+            try
+            {
+                if (!TryCopyCharacterStringBytes(expectedTag, universalTagNumber, rented, true, out bytesRead, out bytesWritten))
+                {
+                    Debug.Fail("TryCopyCharacterStringBytes failed with a precomputed size");
+                    throw new CryptographicException();
+                }
+
+                bool copied = TryCopyCharacterString(
+                    new ReadOnlyMemory<byte>(rented, 0, bytesWritten), 
+                    encoding,
+                    destination,
+                    out charsWritten);
+
+                if (copied)
+                {
+                    _data = _data.Slice(bytesRead);
+                }
+
+                return copied;
+            }
+            finally
+            {
+                Array.Clear(rented, 0, bytesWritten);
+                ArrayPool<byte>.Shared.Return(rented);
+            }
+        }
+
+        /// <summary>
+        /// Gets the source data for a character string under a primitive encoding and tagged as
+        /// the universal class tag for the encoding type.
+        /// </summary>
+        /// <param name="encodingType">The UniversalTagNumber for the string encoding type.</param>
+        /// <param name="contents">The content bytes for the UTF8String payload.</param>
+        /// <returns>
+        ///   <c>true</c> if the character string uses a primitive encoding, <c>false</c> otherwise.
+        /// </returns>
+        /// <exception cref="CryptographicException">
+        ///  <ul>
+        ///   <li>No data remains</li>
+        ///   <li>The tag read does not match the expected tag</li>
+        ///   <li>The length is invalid under the chosen encoding rules</li>
+        ///   <li>A CER encoding was chosen and the primitive content length exceeds the maximum allowed</li>
+        /// </ul>
+        /// </exception>
+        public bool TryGetCharacterStringBytes(UniversalTagNumber encodingType, out ReadOnlyMemory<byte> contents)
+        {
+            return TryGetCharacterStringBytes(new Asn1Tag(encodingType), encodingType, out contents);
+        }
+
+        /// <summary>
+        /// Gets the source data for a character string under a primitive encoding.  The contents
+        /// are not validated as belonging to the requested encoding type.
+        /// </summary>
+        /// <param name="expectedTag">The expected tag</param>
+        /// <param name="encodingType">The UniversalTagNumber for the string encoding type.</param>
+        /// <param name="contents">The content bytes for the UTF8String payload.</param>
+        /// <returns>
+        ///   <c>true</c> if the character string uses a primitive encoding, <c>false</c> otherwise.
+        /// </returns>
+        /// <exception cref="CryptographicException">
+        ///  <ul>
+        ///   <li>No data remains</li>
+        ///   <li>The tag read does not match the expected tag</li>
+        ///   <li>The length is invalid under the chosen encoding rules</li>
+        ///   <li>A CER encoding was chosen and the primitive content length exceeds the maximum allowed</li>
+        /// </ul>
+        /// </exception>
+        /// <exception cref="ArgumentOutOfRangeException">
+        ///   <paramref name="encodingType"/> is not a known character string encoding type.
+        /// </exception>
+        public bool TryGetCharacterStringBytes(
+            Asn1Tag expectedTag,
+            UniversalTagNumber encodingType,
+            out ReadOnlyMemory<byte> contents)
+        {
+            CheckCharacterStringEncodingType(encodingType);
+            return TryGetOctetStringBytes(expectedTag, encodingType, out contents);
+        }
+
+        public bool TryCopyCharacterStringBytes(
+            UniversalTagNumber encodingType,
+            Span<byte> destination,
+            out int bytesWritten)
+        {
+            return TryCopyCharacterStringBytes(
+                new Asn1Tag(encodingType),
+                encodingType,
+                destination,
+                out bytesWritten);
+        }
+
+        public bool TryCopyCharacterStringBytes(
+            Asn1Tag expectedTag,
+            UniversalTagNumber encodingType,
+            Span<byte> destination,
+            out int bytesWritten)
+        {
+            CheckCharacterStringEncodingType(encodingType);
+
+            bool copied = TryCopyCharacterStringBytes(
+                expectedTag,
+                encodingType,
+                destination,
+                true,
+                out int bytesRead,
+                out bytesWritten);
+
+            if (copied)
+            {
+                _data = _data.Slice(bytesRead);
+            }
+
+            return copied;
+        }
+
+        public bool TryCopyCharacterString(
+            UniversalTagNumber encodingType,
+            Span<char> destination,
+            out int charsWritten)
+        {
+            return TryCopyCharacterString(
+                new Asn1Tag(encodingType),
+                encodingType,
+                destination,
+                out charsWritten);
+        }
+
+        public bool TryCopyCharacterString(
+            Asn1Tag expectedTag,
+            UniversalTagNumber encodingType,
+            Span<char> destination,
+            out int charsWritten)
+        {
+            Text.Encoding encoding = AsnCharacterStringEncodings.GetEncoding(encodingType);
+            return TryCopyCharacterString(expectedTag, encodingType, encoding, destination, out charsWritten);
+        }
+
+        public string GetCharacterString(UniversalTagNumber encodingType) =>
+            GetCharacterString(new Asn1Tag(encodingType), encodingType);
+            
+        public string GetCharacterString(Asn1Tag expectedTag, UniversalTagNumber encodingType)
+        {
+            Text.Encoding encoding = AsnCharacterStringEncodings.GetEncoding(encodingType);
+            return GetCharacterString(expectedTag, encodingType, encoding);
+        }
+
+        public AsnReader ReadSequence() => ReadSequence(Asn1Tag.Sequence);
+
+        public AsnReader ReadSequence(Asn1Tag expectedTag)
+        {
+            Asn1Tag tag = ReadTagAndLength(out int ? length, out int headerLength);
+            CheckExpectedTag(tag, expectedTag, UniversalTagNumber.Sequence);
+
+            // T-REC-X.690-201508 sec 8.9.1
+            // T-REC-X.690-201508 sec 8.10.1
+            if (!tag.IsConstructed)
+            {
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+
+            ReadOnlyMemory<byte> contents;
+            int suffix = 0;
+
+            if (length != null)
+            {
+                contents = Slice(_data, headerLength, length.Value);
+            }
+            else
+            {
+                contents = SeekEndOfContents(_data.Slice(headerLength), _ruleSet, 0);
+                suffix = EndOfContentsEncodedLength;
+            }
+
+            _data = _data.Slice(headerLength + contents.Length + suffix);
+            return new AsnReader(contents, _ruleSet);
+        }
+
+        /// <summary>
+        /// Builds a new AsnReader over the bytes bounded by the current position which
+        /// corresponds to an ASN.1 SET OF value, validating the CER or DER sort ordering
+        /// unless suppressed.
+        /// </summary>
+        /// <param name="skipSortOrderValidation">
+        ///   <c>false</c> to validate the sort ordering of the contents, <c>true</c> to
+        ///   allow reading the data without verifying it was properly sorted by the writer.
+        /// </param>
+        /// <returns>An AsnReader over the current position, bounded by the contained length value.</returns>
+        public AsnReader ReadSetOf(bool skipSortOrderValidation = false) =>
+            ReadSetOf(Asn1Tag.SetOf, skipSortOrderValidation);
+
+        public AsnReader ReadSetOf(Asn1Tag expectedTag, bool skipSortOrderValidation = false)
+        {
+            Asn1Tag tag = ReadTagAndLength(out int ? length, out int headerLength);
+            CheckExpectedTag(tag, expectedTag, UniversalTagNumber.SetOf);
+
+            // T-REC-X.690-201508 sec 8.12.1
+            if (!tag.IsConstructed)
+            {
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+
+            ReadOnlyMemory<byte> contents;
+            int suffix = 0;
+
+            if (length != null)
+            {
+                contents = Slice(_data, headerLength, length.Value);
+            }
+            else
+            {
+                contents = SeekEndOfContents(_data.Slice(headerLength), _ruleSet);
+                suffix = EndOfContentsEncodedLength;
+            }
+
+            if (!skipSortOrderValidation)
+            {
+                // T-REC-X.690-201508 sec 11.6
+                // BER data is not required to be sorted.
+                if (_ruleSet == AsnEncodingRules.DER ||
+                    _ruleSet == AsnEncodingRules.CER)
+                {
+                    AsnReader reader = new AsnReader(contents, _ruleSet);
+
+                    ReadOnlySpan<byte> current = ReadOnlySpan<byte>.Empty;
+
+                    while (reader.HasData)
+                    {
+                        ReadOnlySpan<byte> previous = current;
+                        current = reader.GetEncodedValue().Span;
+
+                        int end = Math.Min(previous.Length, current.Length);
+                        int i;
+
+                        for (i = 0; i < end; i++)
+                        {
+                            byte currentVal = current[i];
+                            byte previousVal = previous[i];
+
+                            if (currentVal > previousVal)
+                            {
+                                break;
+                            }
+
+                            if (currentVal < previousVal)
+                            {
+                                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                            }
+                        }
+
+                        if (i == end)
+                        {
+                            // If everything was a tie then we treat the shorter thing as if it were
+                            // followed by an infinite number of 0x00s.  So "previous" better not have
+                            // more data, or if it does, none of it can be non-zero.
+                            //
+                            // Note: It doesn't seem possible for the tiebreaker to matter.
+                            // In DER everything is length prepended, so the content is only compared
+                            // if the tag and length were the same.
+                            //
+                            // In CER you could have an indefinite octet string, but it will contain
+                            // primitive octet strings and EoC. So at some point an EoC is compared
+                            // against a tag, and the sort order is determined.
+                            //
+                            // But since the spec calls it out, maybe there's something degenerate, so
+                            // we'll guard against it anyways.
+
+                            for (; i < previous.Length; i++)
+                            {
+                                if (previous[i] != 0)
+                                {
+                                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+
+            _data = _data.Slice(headerLength + contents.Length + suffix);
+            return new AsnReader(contents, _ruleSet);
+        }
+
+        private static byte GetDigit(byte b)
+        {
+            if (b >= '0' && b <= '9')
+            {
+                return (byte)(b - '0');
+            }
+
+            throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+        }
+
+        private static DateTimeOffset ParseUtcTime(ReadOnlySpan<byte> contentOctets, int twoDigitYearMax)
+        {
+            // The full allowed formats (T-REC-X.680-201510 sec 47.3)
+            // YYMMDDhhmmZ  (a, b1, c1)
+            // YYMMDDhhmm+hhmm (a, b1, c2+)
+            // YYMMDDhhmm-hhmm (a, b1, c2-)
+            // YYMMDDhhmmssZ (a, b2, c1)
+            // YYMMDDhhmmss+hhmm (a, b2, c2+)
+            // YYMMDDhhmmss-hhmm (a, b2, c2-)
+
+            const int AB1C1Length = 11;
+            const int AB1C2Length = AB1C1Length + 4;
+            const int AB2C1Length = AB1C1Length + 2;
+            const int AB2C2Length = AB2C1Length + 4;
+
+            // 11, 13, 15, 17 are legal.
+            // Range check + odd.
+            if (contentOctets.Length < AB1C1Length ||
+                contentOctets.Length > AB2C2Length ||
+                (contentOctets.Length & 1) != 1)
+            {
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+
+            int year = 10 * GetDigit(contentOctets[0]) + GetDigit(contentOctets[1]);
+            int month = 10 * GetDigit(contentOctets[2]) + GetDigit(contentOctets[3]);
+            int day = 10 * GetDigit(contentOctets[4]) + GetDigit(contentOctets[5]);
+            int hour = 10 * GetDigit(contentOctets[6]) + GetDigit(contentOctets[7]);
+            int minute = 10 * GetDigit(contentOctets[8]) + GetDigit(contentOctets[9]);
+            int second = 0;
+            int offsetHour = 0;
+            int offsetMinute = 0;
+            bool minus = false;
+
+            if (contentOctets.Length == AB1C1Length)
+            {
+                if (contentOctets[10] != 'Z')
+                {
+                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                }
+            }
+            else if (contentOctets.Length == AB1C2Length)
+            {
+                if (contentOctets[10] == '-')
+                {
+                    minus = true;
+                }
+                else if (contentOctets[10] != '+')
+                {
+                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                }
+
+                offsetHour = 10 * GetDigit(contentOctets[11]) + GetDigit(contentOctets[12]);
+                offsetMinute = 10 * GetDigit(contentOctets[13]) + GetDigit(contentOctets[14]);
+            }
+            else
+            {
+                second = 10 * GetDigit(contentOctets[10]) + GetDigit(contentOctets[11]);
+
+                if (contentOctets.Length == AB2C1Length)
+                {
+                    if (contentOctets[12] != 'Z')
+                    {
+                        throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                    }
+                }
+                else
+                {
+                    Debug.Assert(contentOctets.Length == AB2C2Length);
+
+                    if (contentOctets[12] == '-')
+                    {
+                        minus = true;
+                    }
+                    else if (contentOctets[12] != '+')
+                    {
+                        throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                    }
+
+                    offsetHour = 10 * GetDigit(contentOctets[13]) + GetDigit(contentOctets[14]);
+                    offsetMinute = 10 * GetDigit(contentOctets[15]) + GetDigit(contentOctets[16]);
+                }
+            }
+
+            TimeSpan offset = new TimeSpan(offsetHour, offsetMinute, 0);
+
+            if (minus)
+            {
+                offset = TimeSpan.Zero - offset;
+            }
+
+            int y = year % 100;
+            int scaledYear = ((twoDigitYearMax / 100 - (y > twoDigitYearMax % 100 ? 1 : 0)) * 100 + y);
+
+            try
+            {
+                return new DateTimeOffset(scaledYear, month, day, hour, minute, second, offset);
+            }
+            catch (Exception e)
+            {
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding, e);
+            }
+        }
+
+        public DateTimeOffset GetUtcTime(int twoDigitYearMax = 2049) =>
+            GetUtcTime(Asn1Tag.UtcTime, twoDigitYearMax);
+
+        /// <summary>
+        /// Gets the DateTimeOffset represented by a UTCTime value.
+        /// </summary>
+        /// <param name="expectedTag">The expected tag</param>
+        /// <param name="twoDigitYearMax">
+        /// The largest year to represent with this value.
+        /// The default value, 2049, represents the 1950-2049 range for X.509 certificates.
+        /// </param>
+        /// <returns>
+        /// A DateTimeOffset representing the value encoded in the UTCTime.
+        /// </returns>
+        /// <seealso cref="System.Globalization.Calendar.TwoDigitYearMax"/>
+        public DateTimeOffset GetUtcTime(Asn1Tag expectedTag, int twoDigitYearMax = 2049)
+        {
+            // T-REC-X.680-201510 sec 47.3 says it is IMPLICIT VisibleString, which means
+            // that BER is allowed to do complex constructed forms.
+
+            // The full allowed formats (T-REC-X.680-201510 sec 47.3)
+            // YYMMDDhhmmZ  (a, b1, c1)
+            // YYMMDDhhmm+hhmm (a, b1, c2+)
+            // YYMMDDhhmm-hhmm (a, b1, c2-)
+            // YYMMDDhhmmssZ (a, b2, c1)
+            // YYMMDDhhmmss+hhmm (a, b2, c2+)
+            // YYMMDDhhmmss-hhmm (a, b2, c2-)
+
+            // CER and DER are restricted to YYMMDDhhmmssZ
+            // T-REC-X.690-201510 sec 11.8
+            
+            // Optimize for the CER/DER primitive encoding:
+            if (TryGetOctetStringBytes(
+                expectedTag,
+                out ReadOnlyMemory<byte> primitiveOctets,
+                out int headerLength,
+                UniversalTagNumber.UtcTime))
+            {
+                if (primitiveOctets.Length == 13)
+                {
+                    DateTimeOffset value = ParseUtcTime(primitiveOctets.Span, twoDigitYearMax);
+                    _data = _data.Slice(headerLength + primitiveOctets.Length);
+                    return value;
+                }
+            }
+
+            // T-REC-X.690-201510 sec 11.8
+            if (_ruleSet == AsnEncodingRules.DER || _ruleSet == AsnEncodingRules.CER)
+            {
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+
+            // The longest legal format is (a, b2, c2), which comes out to 17 characters/bytes.
+            byte[] rented = ArrayPool<byte>.Shared.Rent(17);
+            ReadOnlySpan<byte> contentOctets = ReadOnlySpan<byte>.Empty;
+
+            try
+            {
+                if (TryCopyCharacterStringBytes(
+                    expectedTag,
+                    UniversalTagNumber.UtcTime,
+                    rented,
+                    true,
+                    out int bytesRead,
+                    out int contentLength))
+                {
+                    contentOctets = Slice(rented, 0, contentLength);
+
+                    DateTimeOffset value = ParseUtcTime(contentOctets, twoDigitYearMax);
+                    // Includes the header
+                    _data = _data.Slice(bytesRead);
+                    return value;
+                }
+            }
+            finally
+            {
+                Array.Clear(rented, 0, contentOctets.Length);
+                ArrayPool<byte>.Shared.Return(rented);
+            }
+
+            throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+        }
+
+        private static DateTimeOffset ParseGeneralizedTime(
+            AsnEncodingRules ruleSet,
+            ReadOnlySpan<byte> contentOctets,
+            bool disallowFractions)
+        {
+            // T-REC-X.680-201510 sec 46 defines a lot of formats for GeneralizedTime.
+            //
+            // All formats start with yyyyMMdd.
+            //
+            // "Local time" formats are
+            //   [date]HH.fractionOfAnHourToAnArbitraryPrecision
+            //   [date]HHmm.fractionOfAMinuteToAnArbitraryPrecision
+            //   [date]HHmmss.fractionOfASecondToAnArbitraryPrecision
+            //
+            // "UTC time" formats are the local formats suffixed with 'Z'
+            //
+            // "UTC offset time" formats are the local formats suffixed with
+            //  +HH
+            //  +HHmm
+            //  -HH
+            //  -HHmm
+            // Additionally, it us unclear if the following formats are supposed to be supported,
+            // because the ISO 8601:2004 spec is behind a paywall.
+            //  +HH:mm
+            //  -HH:mm
+            //
+            // Also, every instance of '.' is actually "period or comma".
+
+            // Since DateTimeOffset doesn't have a notion of
+            // "I'm a local time, but with an unknown offset", the computer's current offset will
+            // be used.
+
+            // T-REC-X.690-201510 sec 11.7 binds CER and DER to a much smaller set of inputs:
+            //  * Only the UTC/Z format can be used.
+            //  * HHmmss must always be used
+            //  * If fractions are present they will be separated by period, never comma.
+            //  * If fractions are present the last digit mustn't be 0.
+
+            bool strict = ruleSet == AsnEncodingRules.DER || ruleSet == AsnEncodingRules.CER;
+            if (strict && contentOctets.Length < 15)
+            {
+                // yyyyMMddHHmmssZ
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+            else if (contentOctets.Length < 10)
+            {
+                // yyyyMMddHH
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+
+            int offset = 0;
+            int year =
+                1000 * GetDigit(contentOctets[offset++]) +
+                100 * GetDigit(contentOctets[offset++]) +
+                10 * GetDigit(contentOctets[offset++]) +
+                GetDigit(contentOctets[offset++]);
+
+            int month = 10 * GetDigit(contentOctets[offset++]) + GetDigit(contentOctets[offset++]);
+            int day = 10 * GetDigit(contentOctets[offset++]) + GetDigit(contentOctets[offset++]);
+            int hour = 10 * GetDigit(contentOctets[offset++]) + GetDigit(contentOctets[offset++]);
+            int? minute = null;
+            int? second = null;
+            ulong fraction = 0;
+            ulong fractionScale = 1;
+            TimeSpan? timeOffset = null;
+            bool isZulu = false;
+
+            const byte HmsState = 0;
+            const byte FracState = 1;
+            const byte SuffixState = 2;
+            byte state = HmsState;
+            
+            if (contentOctets.Length > offset)
+            {
+                byte octet = contentOctets[offset];
+
+                if (octet == 'Z' || octet == '-' || octet == '+')
+                {
+                    state = SuffixState;
+                }
+                else if (octet == '.' || octet == ',')
+                {
+                    state = FracState;
+                }
+                else if (contentOctets.Length - 1 <= offset)
+                {
+                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                }
+                else
+                {
+                    minute = 10 * GetDigit(contentOctets[offset++]) + GetDigit(contentOctets[offset++]);
+                }
+            }
+
+            if (state == HmsState && contentOctets.Length > offset)
+            {
+                byte octet = contentOctets[offset];
+
+                if (octet == 'Z' || octet == '-' || octet == '+')
+                {
+                    state = SuffixState;
+                }
+                else if (octet == '.' || octet == ',')
+                {
+                    state = FracState;
+                }
+                else if (contentOctets.Length - 1 <= offset)
+                {
+                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                }
+                else
+                {
+                    second = 10 * GetDigit(contentOctets[offset++]) + GetDigit(contentOctets[offset++]);
+                }
+            }
+
+            if (state == HmsState && contentOctets.Length > offset)
+            {
+                byte octet = contentOctets[offset];
+
+                if (octet == 'Z' || octet == '-' || octet == '+')
+                {
+                    state = SuffixState;
+                }
+                else if (octet == '.' || octet == ',')
+                {
+                    state = FracState;
+                }
+                else
+                {
+                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                }
+            }
+
+            if (state == FracState)
+            {
+                if (disallowFractions)
+                {
+                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                }
+
+                Debug.Assert(contentOctets.Length > offset);
+                byte octet = contentOctets[offset++];
+
+                if (octet == '.')
+                {
+                    // Always valid
+                }
+                else if (octet == ',')
+                {
+                    // Valid for BER, but not CER or DER.
+                    if (strict)
+                    {
+                        throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                    }
+                }
+                else
+                {
+                    Debug.Fail($"Unhandled value '{octet:X2}' in {nameof(FracState)}");
+                    throw new CryptographicException();
+                }
+
+                // There are 36,000,000,000 ticks per hour, and hour is our largest scale.
+                // In case the double -> Ticks conversion allows for rounding up we can allow
+                // for a 12th digit.
+                const ulong MaxScale = 1_000_000_000_000;
+
+                if (contentOctets.Length <= offset)
+                {
+                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                }
+
+                for (; offset < contentOctets.Length; offset++)
+                {
+                    octet = contentOctets[offset];
+
+                    if (octet == 'Z' || octet == '-' || octet == '+')
+                    {
+                        state = SuffixState;
+                        break;
+                    }
+
+                    if (fractionScale < MaxScale)
+                    {
+                        fraction *= 10;
+                        fraction += GetDigit(contentOctets[offset]);
+                        fractionScale *= 10;
+                    }
+                    else
+                    {
+                        GetDigit(contentOctets[offset]);
+                    }
+                }
+            }
+
+            if (state == SuffixState)
+            {
+                Debug.Assert(contentOctets.Length > offset);
+                byte octet = contentOctets[offset++];
+
+                if (octet == 'Z')
+                {
+                    timeOffset = TimeSpan.Zero;
+                    isZulu = true;
+                }
+                else
+                {
+                    bool isMinus;
+
+                    if (octet == '+')
+                    {
+                        isMinus = false;
+                    }
+                    else if (octet == '-')
+                    {
+                        isMinus = true;
+                    }
+                    else
+                    {
+                        Debug.Fail($"Unhandled value '{octet:X2}' in {nameof(SuffixState)}");
+                        throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                    }
+
+                    if (contentOctets.Length - 1 <= offset)
+                    {
+                        throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                    }
+
+                    int offsetHour = 10 * GetDigit(contentOctets[offset++]) + GetDigit(contentOctets[offset++]);
+                    int offsetMinute = 0;
+
+                    if (contentOctets.Length > offset)
+                    {
+                        if (contentOctets[offset] == ':')
+                        {
+                            offset++;
+                        }
+                    }
+
+                    if (contentOctets.Length - 1 > offset)
+                    {
+                        offsetMinute = 10 * GetDigit(contentOctets[offset++]) + GetDigit(contentOctets[offset++]);
+                    }
+
+                    TimeSpan tmp = new TimeSpan(offsetHour, offsetMinute, 0);
+
+                    if (isMinus)
+                    {
+                        tmp = TimeSpan.Zero - tmp;
+                    }
+
+                    timeOffset = tmp;
+                }
+            }
+
+            // Was there data after a suffix?
+            if (offset != contentOctets.Length)
+            {
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+
+            // T-REC-X.690-201510 sec 11.7
+            if (strict)
+            {
+                if (!isZulu || !second.HasValue)
+                {
+                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                }
+
+                if (fraction != 0 && fraction % 10 == 0)
+                {
+                    throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+                }
+            }
+
+            double frac = (double)fraction / fractionScale;
+            TimeSpan fractionSpan = TimeSpan.Zero;
+
+            if (!minute.HasValue)
+            {
+                minute = 0;
+                second = 0;
+
+                if (fraction != 0)
+                {
+                    // No minutes means this is fractions of an hour
+                    fractionSpan = TimeSpan.FromHours(frac);
+                }
+            }
+            else if (!second.HasValue)
+            {
+                second = 0;
+
+                if (fraction != 0)
+                {
+                    // No seconds means this is fractions of a minute
+                    fractionSpan = TimeSpan.FromMinutes(frac);
+                }
+            }
+            else if (fraction != 0)
+            {
+                // Both minutes and seconds means fractions of a second.
+                fractionSpan = TimeSpan.FromSeconds(frac);
+            }
+            
+            DateTimeOffset value;
+
+            if (timeOffset == null)
+            {
+                value = new DateTimeOffset(new DateTime(year, month, day, hour, minute.Value, second.Value));
+            }
+            else
+            {
+                value = new DateTimeOffset(year, month, day, hour, minute.Value, second.Value, timeOffset.Value);
+            }
+
+            value += fractionSpan;
+            return value;
+        }
+
+        public DateTimeOffset GetGeneralizedTime(bool disallowFractions=false) =>
+            GetGeneralizedTime(Asn1Tag.GeneralizedTime, disallowFractions);
+
+        public DateTimeOffset GetGeneralizedTime(Asn1Tag expectedTag, bool disallowFractions=false)
+        {
+            if (TryGetOctetStringBytes(
+                expectedTag,
+                out ReadOnlyMemory<byte> primitiveOctets,
+                out int headerLength,
+                UniversalTagNumber.GeneralizedTime))
+            {
+                DateTimeOffset value = ParseGeneralizedTime(_ruleSet, primitiveOctets.Span, disallowFractions);
+                _data = _data.Slice(headerLength + primitiveOctets.Length);
+                return value;
+            }
+
+            // T-REC-X.690-201510 sec 9.2
+            // T-REC-X.690-201510 sec 10.2
+            if (_ruleSet == AsnEncodingRules.DER || _ruleSet == AsnEncodingRules.CER)
+            {
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+
+            int upperBound = PeekContentBytes().Length;
+
+            byte[] rented = ArrayPool<byte>.Shared.Rent(upperBound);
+            ReadOnlySpan<byte> contentOctets = ReadOnlySpan<byte>.Empty;
+
+            try
+            {
+                if (TryCopyCharacterStringBytes(
+                    expectedTag,
+                    UniversalTagNumber.GeneralizedTime,
+                    rented,
+                    true,
+                    out int bytesRead,
+                    out int contentLength))
+                {
+                    contentOctets = Slice(rented, 0, contentLength);
+
+                    DateTimeOffset value = ParseGeneralizedTime(_ruleSet, contentOctets, disallowFractions);
+                    // Includes the header
+                    _data = _data.Slice(bytesRead);
+                    return value;
+                }
+            }
+            finally
+            {
+                Array.Clear(rented, 0, contentOctets.Length);
+                ArrayPool<byte>.Shared.Return(rented);
+            }
+
+            throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+        }
+
+        private static ReadOnlySpan<byte> Slice(ReadOnlySpan<byte> source, int offset, int length)
+        {
+            Debug.Assert(offset >= 0);
+
+            if (length < 0 || source.Length - offset < length)
+            {
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+
+            return source.Slice(offset, length);
+        }
+
+        private static ReadOnlyMemory<byte> Slice(ReadOnlyMemory<byte> source, int offset, int? length)
+        {
+            Debug.Assert(offset >= 0);
+
+            if (length == null)
+            {
+                return source.Slice(offset);
+            }
+
+            int lengthVal = length.Value;
+
+            if (lengthVal < 0 || source.Length - offset < lengthVal)
+            {
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+
+            return source.Slice(offset, lengthVal);
+        }
+
+        private static void CheckEncodingRules(AsnEncodingRules ruleSet)
+        {
+            if (ruleSet != AsnEncodingRules.BER &&
+                ruleSet != AsnEncodingRules.CER &&
+                ruleSet != AsnEncodingRules.DER)
+            {
+                throw new ArgumentOutOfRangeException(nameof(ruleSet));
+            }
+        }
+
+        private static void CheckExpectedTag(Asn1Tag tag, Asn1Tag expectedTag, UniversalTagNumber tagNumber)
+        {
+            if (expectedTag.TagClass == TagClass.Universal && expectedTag.TagValue != (int)tagNumber)
+            {
+                throw new ArgumentException(
+                    SR.Cryptography_Asn_UniversalValueIsFixed,
+                    nameof(expectedTag));
+            }
+
+            if (expectedTag.TagClass != tag.TagClass || expectedTag.TagValue != tag.TagValue)
+            {
+                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
+            }
+        }
+
+        private static void CheckCharacterStringEncodingType(UniversalTagNumber encodingType)
+        {
+            // T-REC-X.680-201508 sec 41
+            switch (encodingType)
+            {
+                case UniversalTagNumber.BMPString:
+                case UniversalTagNumber.GeneralString:
+                case UniversalTagNumber.GraphicString:
+                case UniversalTagNumber.IA5String:
+                case UniversalTagNumber.ISO646String:
+                case UniversalTagNumber.NumericString:
+                case UniversalTagNumber.PrintableString:
+                case UniversalTagNumber.TeletexString:
+                // T61String is an alias for TeletexString (already listed)
+                case UniversalTagNumber.UniversalString:
+                case UniversalTagNumber.UTF8String:
+                case UniversalTagNumber.VideotexString:
+                    // VisibleString is an alias for ISO646String (already listed)
+                    return;
+            }
+
+            throw new ArgumentOutOfRangeException(nameof(encodingType));
+        }
+    }
+}
